backend/requirements.txt
fastapi==0.104.1
uvicorn==0.40.0
pydantic==2.5.2
python-multipart==0.0.9
sqlalchemy>=2.0.0
python-dotenv
slowapi
PyYAML
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
apscheduler>=3.10.0
httpx==0.27.0
bcrypt==4.1.2
passlib==1.7.4
PyJWT>=2.8.0
python-jose>=3.3.0
Pillow>=10.0.0
fpdf2>=2.7.0
cryptography>=41.0.0
python-json-logger>=2.0.0
sentry-sdk>=1.40.0
psutil>=5.9.0
prometheus-client>=0.18.0
redis>=5.0.0
reportlab>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
